backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
boto3==1.29.0
strands-agents==0.1.0
strands-agents-tools==0.1.0
thefuzz==0.20.0
python-Levenshtein==0.23.0
aiosmtplib==3.0.1
email-validator==2.1.0
